+++ java-client-sdk/lib/src/main/java/industries/dot/brease/SDKConfiguration.java
@@ -13,7 +13,7 @@ class SDKConfiguration {
 	public String serverUrl;
 	public int serverIdx = 0;
 	public String language = "java";
-	public String sdkVersion = "0.3.0";
-	public String genVersion = "2.35.3";
+	public String sdkVersion = "0.2.3";
+	public String genVersion = "2.35.9";
 	
 }